Implemented Tmux layout

+++ src/terminal/mod.rs
@@ -190,10 +190,10 @@ impl Terminal {
         (cols, rows)
     }
 
-    pub fn get_char_width_height(&self) -> (i64, i64) {
+    pub fn get_char_width_height(&self) -> (i32, i32) {
         let binding = self.imp().vte.borrow();
         let vte = binding.as_ref().unwrap();
-        (vte.char_width(), vte.char_height())
+        (vte.char_width() as i32, vte.char_height() as i32)
     }
 }
